requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy
pandas
scikit-learn
db-dtypes

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-storage
google-cloud-bigquery
mlflow
s3fs

# utilities
six>=1.14
joblib
memoized-property
termcolor
colorama
retrying
